fastapi_user_mgmt/app/database/session.py
from motor.motor_asyncio import AsyncIOMotorClient
from dotenv import load_dotenv
import os

load_dotenv()

MONGODB_URL = os.getenv("MONGODB_URL")
MONGODB_DB = os.getenv("MONGODB_DB")

client = AsyncIOMotorClient(MONGODB_URL)
db = client[MONGODB_DB]