requirements.txt
fastapi
motor
pydantic
python-dotenv
uvicorn
passlib
bcrypt
python-jose[cryptography]